+++ packages/oidc/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @batoanng/oidc
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/mui-components@3.0.22
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/mui-components@3.0.21
+
 ## 0.0.17
 
 ### Patch Changes